# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ForecastRequest": {
        "additionalProperties": true,
        "properties": {
          "frame": {
            "anyOf": [
              {
                "maximum": 29.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alternative alias for frame_id",
            "examples": [
              5
            ],
            "title": "Frame"
          },
          "frame_id": {
            "anyOf": [
              {
                "maximum": 29.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame ID 0..29 inclusive (422 if out of range)",
            "examples": [
              5
            ],
            "title": "Frame Id"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alternative alias for scenario_id",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Scenario ID - must exist (404 if not)",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario Id"
          }
        },
        "title": "ForecastRequest",
        "type": "object"
      },
      "ForecastResponse": {
        "additionalProperties": true,
        "properties": {
          "edges": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NetworkEdge"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Edges observed",
            "title": "Edges"
          },
          "engine_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Engine metadata"
          },
          "flow_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/FlowSummary"
              },
              {
                "type": "null"
              }
            ],
            "description": "Flow summary observed + model-predicted"
          },
          "forecast": {
            "allOf": [
              {
                "$ref": "#/components/schemas/Forecast"
              }
            ],
            "description": "Forecast with raw/smoothed/slope/stage/ranking/evidence/MITRE/engine metadata"
          },
          "frame_id": {
            "description": "Frame ID",
            "examples": [
              5
            ],
            "title": "Frame Id",
            "type": "integer"
          },
          "groundTruth": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/GroundTruth"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias camelCase"
          },
          "ground_truth": {
            "allOf": [
              {
                "$ref": "#/components/schemas/GroundTruth"
              }
            ],
            "description": "Gated ground truth - revealed false hides label_distribution before correct frame"
          },
          "nodes": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/NetworkNode"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Nodes observed",
            "title": "Nodes"
          },
          "scenario_id": {
            "description": "Scenario ID",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario Id",
            "type": "string"
          },
          "signals": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Signals"
              },
              {
                "type": "null"
              }
            ],
            "description": "Signals raw/smoothed/slope/warning"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Timestamp",
            "title": "Timestamp"
          }
        },
        "required": [
          "scenario_id",
          "frame_id",
          "forecast",
          "ground_truth"
        ],
        "title": "ForecastResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NetworkEdge": {
        "additionalProperties": true,
        "properties": {
          "activity": {
            "description": "Observed activity count in frame (observed)",
            "examples": [
              3
            ],
            "title": "Activity",
            "type": "integer"
          },
          "first_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame first seen",
            "examples": [
              0
            ],
            "title": "First Seen"
          },
          "last_seen": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame last seen",
            "examples": [
              5
            ],
            "title": "Last Seen"
          },
          "novelty": {
            "description": "Edge novelty 0..1 graph-derived",
            "examples": [
              0.8
            ],
            "title": "Novelty",
            "type": "number"
          },
          "protocol": {
            "description": "Protocol TCP|UDP|ICMP",
            "examples": [
              "TCP"
            ],
            "title": "Protocol",
            "type": "string"
          },
          "protocols": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "Protocols list",
            "examples": [
              [
                "TCP"
              ]
            ],
            "title": "Protocols"
          },
          "source": {
            "description": "Source host alias (observed)",
            "examples": [
              "host-1"
            ],
            "title": "Source",
            "type": "string"
          },
          "status": {
            "description": "Status - observed for real, predicted for dashed predicted path, simulated/removed for isolation",
            "examples": [
              "observed"
            ],
            "title": "Status",
            "type": "string"
          },
          "target": {
            "description": "Target host alias (observed)",
            "examples": [
              "host-2"
            ],
            "title": "Target",
            "type": "string"
          }
        },
        "required": [
          "source",
          "target",
          "protocol",
          "activity",
          "novelty",
          "status"
        ],
        "title": "NetworkEdge",
        "type": "object"
      },
...
      "Signals": {
        "additionalProperties": true,
        "properties": {
          "alpha": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": 0.4,
            "description": "EWMA alpha",
            "examples": [
              0.4
            ],
            "title": "Alpha"
          },
          "raw_risk": {
            "description": "Raw risk probability - learned (Random Forest)",
            "examples": [
              0.22
            ],
            "title": "Raw Risk",
            "type": "number"
          },
          "slope": {
            "description": "Five-frame risk slope - rule-derived",
            "examples": [
              0.015
            ],
            "title": "Slope",
            "type": "number"
          },
          "smoothed_risk": {
            "description": "Five-frame EWMA smoothed risk (alpha 0.4) - rule-derived",
            "examples": [
              0.25
            ],
            "title": "Smoothed Risk",
            "type": "number"
          },
          "threshold": {
            "description": "Frozen warning threshold from validation",
            "examples": [
              0.45
            ],
            "title": "Threshold",
            "type": "number"
          },
          "warning": {
            "description": "Early warning requires smoothed>threshold and positive slope",
            "examples": [
              false
            ],
            "title": "Warning",
            "type": "boolean"
          }
        },
        "required": [
          "raw_risk",
          "smoothed_risk",
          "slope",
          "warning",
          "threshold"
        ],
        "title": "Signals",
        "type": "object"
      },
...
      "SimulationRequest": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "description": "Action must be isolate_host (422 if unsupported)",
            "examples": [
              "isolate_host"
            ],
            "title": "Action",
            "type": "string"
          },
          "frame": {
            "anyOf": [
              {
                "maximum": 29.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias for frame_id",
            "examples": [
              10
            ],
            "title": "Frame"
          },
          "frame_id": {
            "anyOf": [
              {
                "maximum": 29.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Frame ID 0..29 (422 if out of range)",
            "examples": [
              10
            ],
            "title": "Frame Id"
          },
          "host": {
            "description": "Host to isolate - must exist in frame nodes (422 if not)",
            "examples": [
              "host-42"
            ],
            "title": "Host",
            "type": "string"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Alias for scenario_id",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario"
          },
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Scenario ID must exist (404 if not)",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario Id"
          }
        },
        "required": [
          "host",
          "action"
        ],
        "title": "SimulationRequest",
        "type": "object"
      },
      "SimulationResult": {
        "additionalProperties": true,
        "properties": {
          "action": {
            "description": "Action performed isolate_host",
            "examples": [
              "isolate_host"
            ],
            "title": "Action",
            "type": "string"
          },
          "after": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SimulationBeforeAfter"
              }
            ],
            "description": "After isolation forecast - simulated state orange for simulation, muted edge for removed simulated path"
          },
          "before": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SimulationBeforeAfter"
              }
            ],
            "description": "Before isolation forecast - observed/predicted state"
          },
          "deltas": {
            "allOf": [
              {
                "$ref": "#/components/schemas/SimulationDeltas"
              }
            ],
            "description": "Deltas before/after"
          },
          "engine_metadata": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EngineMetadata"
              },
              {
                "type": "null"
              }
            ],
            "description": "Engine metadata"
          },
          "frame_id": {
            "description": "Frame ID",
            "examples": [
              10
            ],
            "title": "Frame Id",
            "type": "integer"
          },
          "host": {
            "description": "Host isolated (observed state) - host alias safe, never raw IP",
            "examples": [
              "host-42"
            ],
            "title": "Host",
            "type": "string"
          },
          "label": {
            "description": "Label must be 'Estimated simulated effect - not causal proof'",
            "examples": [
              "Estimated simulated effect - not causal proof"
            ],
            "title": "Label",
            "type": "string"
          },
          "limitations": {
            "description": "Limitations must contain 'Estimated simulated effect - not causal proof'",
            "examples": [
              "Estimated simulated effect - not causal proof. Binary benign-versus-malicious risk only; isolation is simulated estimate - not causal proof; recalculated graph features and risk inputs via same pipeline. Original replay unchanged."
            ],
            "title": "Limitations",
            "type": "string"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Message alias for label",
            "examples": [
              "Estimated simulated effect - not causal proof"
            ],
            "title": "Message"
          },
          "original_unchanged": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "default": true,
            "description": "Whether original replay remains unchanged after simulation (immutable)",
            "title": "Original Unchanged"
          },
          "removed_edges": {
            "description": "Removed or down-weighted suspicious host active edges - muted edge style",
            "items": {
              "$ref": "#/components/schemas/NetworkEdge"
            },
            "title": "Removed Edges",
            "type": "array"
          },
          "scenario_id": {
            "description": "Scenario ID",
            "examples": [
              "cyberworld-replay-v1"
            ],
            "title": "Scenario Id",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "frame_id",
          "host",
          "action",
          "before",
          "after",
          "removed_edges",
          "deltas",
          "limitations",
          "label"
        ],
        "title": "SimulationResult",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Predictive cybersecurity decision-support prototype - Replay -> Risk model -> Temporal state -> Stage estimate -> Target ranking -> Evidence and MITRE -> Simulated isolation -> Before/after comparison",
    "title": "CyberWorld AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/forecast": {
      "post": {
        "description": "POST /forecast - input {scenario_id, frame_id} or {scenario, frame}.\n\nReturns Forecast with raw risk, smoothed risk, slope, stage estimate,\ntarget ranking, evidence, MITRE, engine metadata.\n\nEnforces:\n- scenario_id must exist (404 if not)\n- frame_id must be within 0..29 (422 if out of range, handled by Pydantic)\n- Test labels are not returned before their frame: ground_truth.revealed false with no label_distribution\n  (ensured via stored scenario gating, and we double-check to strip any leaked fields)\n- Future labels must not be exposed to forecasting pipeline (already handled in scenario.py)",
        "operationId": "post_forecast_api_v1_forecast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get forecast for a frame",
        "tags": [
          "forecast"
        ]
      }
    },
...
    "/api/v1/simulate/isolate-host": {
      "post": {
        "description": "POST /simulate/isolate-host - clones the selected frame, removes or down-weights\nsuspicious host's active edges, recalculates graph features and risk inputs,\nre-runs same pipeline, returns SimulationResult before/after, removed edges,\ndeltas, limitations, label 'Estimated simulated effect - not causal proof'.\n\nValidates host exists and action supported. Original replay unchanged.",
        "operationId": "simulate_isolate_api_v1_simulate_isolate_host_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Isolate suspicious host (simulated)",
        "tags": [
          "simulation"
        ]
      }
    }
  }
}